Merge pull request n64decomp#36 from ahouts/master

+++ include/PR/os_misc.h
@@ -1,11 +1,11 @@
 #ifndef _ULTRA64_OS_MISC_H_
 #define _ULTRA64_OS_MISC_H_
-
+#include <PR/ultratypes.h>
 /* Miscellaneous OS functions */
 
 void osInitialize(void);
 u32 osGetCount(void);
 
-u32 osVirtualToPhysical(void *);
+uintptr_t osVirtualToPhysical(void *);
 
 #endif
